// src/api-generated/model/updateProductDto.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * SaaS API
 * The core API for my multi-tenant SaaS
 * OpenAPI spec version: 1.0
 */
import type { UpdateProductDtoType } from './updateProductDtoType';
import type { UpdateProductDtoSpecifications } from './updateProductDtoSpecifications';
import type { ProductAddonDto } from './productAddonDto';
import type { ProductMediaDto } from './productMediaDto';
import type { ProductInstanceDto } from './productInstanceDto';

export interface UpdateProductDto {
  type?: UpdateProductDtoType;
  title?: string;
  description?: string;
  base_price?: number;
  currency?: string;
  specifications?: UpdateProductDtoSpecifications;
  available_addons?: ProductAddonDto[];
  media?: ProductMediaDto[];
  instances?: ProductInstanceDto[];
}